Remove Contributing and Testing Documentation, and Delete Unused Store and Component Files
- Deleted the CONTRIBUTING.md and TESTING.md files to streamline documentation.
- Removed unused Svelte store (stores.ts) and component files (EnhancedImage.svelte, ResourceTree.svelte) to clean up the codebase.
- Updated resource type definitions in resource.ts to reflect current usage.

These changes enhance project maintainability and align with best practices for documentation and code organization.
