+++ src/lib/types/index.ts
@@ -1,0 +1,4 @@
+export type { EnhancedImage } from './image';
+export type { Video, YouTubeChannel } from './video';
+export type { Project } from './project';
+export type { Resource, TreeNode } from './resource';
