+++ src/lib/types/project.ts
@@ -1,0 +1,7 @@
+export interface Project {
+	title: string;
+	description: string;
+	tech: string[];
+	url: string;
+	status: 'active' | 'stable' | 'experimental';
+}
